requirements.txt
fastapi==0.104.1
ultralytics>=8.3.0
pillow==10.1.0
python-multipart==0.0.6
